refactor(settings): put the tab on the shared controls and split the sections
Half the settings tab was written in a second style from the other half. The
transcription section used the declarative ctx builders; everything else -
recording, formats, storage, splitting, multi-track, the player, input
processing, cleanup defaults, diagnostics - hand-rolled `new Setting(...)`
rows with their own `await saveSettings()` in each onChange: 47 rows and 36
save calls, none of which got the builders' debounced text saves, disabled
rendering, or help links.

Add sectionContext() and move every row the declarative model covers onto the
shared builders: 36 save calls drop to 15 and 47 raw rows to 22. What remains
imperative is what genuinely does not fit, now stated in the module comment -
live device enumeration, the async format-availability probe, the output
summary, the Map-driven per-track rows, the part-suffix field's own validation
feedback, and the diagnostics actions.

The audio-cleanup stage rows (toggle plus one numeric parameter) were written
twice, in the tab and in the cleanup dialog, and only the dialog greyed out the
input of a disabled stage. They now share addStageRowTo, so the two cannot
drift.

The 801-line transcription section splits by subject into the engine fields,
the LLM vendor fields, and the profile managers, leaving a 384-line section
that reads as an outline of the feature.
