+++ src/cleanup/AudioProcessingModal.ts
@@ -8,7 +8,7 @@
 
 import { App, ButtonComponent, Notice, Setting, TFile } from 'obsidian';
 import { PluginModal } from '../ui/PluginModal';
-import { addNumberInputTo } from '../settings/settingControls';
+import { addStageRowTo } from '../settings/settingControls';
 import {
 	MIN_CLEANUP_HIGHPASS_HZ,
 	MAX_CLEANUP_HIGHPASS_HZ,
@@ -72,48 +72,57 @@ export class AudioProcessingModal extends PluginModal {
 			text: 'Produces a processed WAV copy next to the source.',
 		});
 
-		this.addStageWithNumber(
-			contentEl,
-			'High-pass filter',
-			'Remove low-frequency rumble below the cutoff (Hz).',
-			this.config.highPass.enabled,
-			(v) => (this.config.highPass.enabled = v),
-			{
+		addStageRowTo(contentEl, {
+			name: 'High-pass filter',
+			desc: 'Remove low-frequency rumble below the cutoff (Hz).',
+			getEnabled: () => this.config.highPass.enabled,
+			setEnabled: (v) => {
+				this.config.highPass.enabled = v;
+			},
+			value: {
 				min: MIN_CLEANUP_HIGHPASS_HZ,
 				max: MAX_CLEANUP_HIGHPASS_HZ,
 				step: CLEANUP_HIGHPASS_STEP_HZ,
-				value: this.config.highPass.hz,
-				onChange: (v) => (this.config.highPass.hz = v),
+				get: () => this.config.highPass.hz,
+				set: (v) => {
+					this.config.highPass.hz = v;
+				},
 			},
-		);
-		this.addStageWithNumber(
-			contentEl,
-			'Noise gate',
-			'Silence the signal below the threshold (dBFS).',
-			this.config.gate.enabled,
-			(v) => (this.config.gate.enabled = v),
-			{
+		});
+		addStageRowTo(contentEl, {
+			name: 'Noise gate',
+			desc: 'Silence the signal below the threshold (dBFS).',
+			getEnabled: () => this.config.gate.enabled,
+			setEnabled: (v) => {
+				this.config.gate.enabled = v;
+			},
+			value: {
 				min: MIN_CLEANUP_GATE_THRESHOLD_DB,
 				max: MAX_CLEANUP_GATE_THRESHOLD_DB,
 				step: CLEANUP_GATE_STEP_DB,
-				value: this.config.gate.thresholdDb,
-				onChange: (v) => (this.config.gate.thresholdDb = v),
+				get: () => this.config.gate.thresholdDb,
+				set: (v) => {
+					this.config.gate.thresholdDb = v;
+				},
 			},
-		);
-		this.addStageWithNumber(
-			contentEl,
-			'Loudness leveling',
-			'Even out quiet and loud passages; makeup gain (dB).',
-			this.config.leveling.enabled,
-			(v) => (this.config.leveling.enabled = v),
-			{
+		});
+		addStageRowTo(contentEl, {
+			name: 'Loudness leveling',
+			desc: 'Even out quiet and loud passages; makeup gain (dB).',
+			getEnabled: () => this.config.leveling.enabled,
+			setEnabled: (v) => {
+				this.config.leveling.enabled = v;
+			},
+			value: {
 				min: MIN_CLEANUP_LEVELING_MAKEUP_DB,
 				max: MAX_CLEANUP_LEVELING_MAKEUP_DB,
 				step: CLEANUP_LEVELING_STEP_DB,
-				value: this.config.leveling.makeupDb,
-				onChange: (v) => (this.config.leveling.makeupDb = v),
+				get: () => this.config.leveling.makeupDb,
+				set: (v) => {
+					this.config.leveling.makeupDb = v;
+				},
 			},
-		);
+		});
 
 		new Setting(contentEl)
 			.setName('Channels')
@@ -167,42 +176,6 @@ export class AudioProcessingModal extends PluginModal {
 					this.close();
 				},
 			},
-		);
-	}
-
-	/**
-	 * Adds a stage toggle with a parameter number input on the same row. The
-	 * input is greyed out while the stage is off so it is clear the parameter
-	 * only takes effect once the stage is enabled.
-	 */
-	private addStageWithNumber(
-		container: HTMLElement,
-		name: string,
-		desc: string,
-		enabled: boolean,
-		onToggle: (value: boolean) => void,
-		field: {
-			min: number;
-			max: number;
-			step: number;
-			value: number;
-			onChange: (value: number) => void;
-		},
-	): void {
-		const setting = new Setting(container).setName(name).setDesc(desc);
-		const numberInput = addNumberInputTo(setting, {
-			min: field.min,
-			max: field.max,
-			step: field.step,
-			get: () => field.value,
-			set: field.onChange,
-		});
-		numberInput.setDisabled(!enabled);
-		setting.addToggle((toggle) =>
-			toggle.setValue(enabled).onChange((value) => {
-				onToggle(value);
-				numberInput.setDisabled(!value);
-			}),
 		);
 	}
